REVO i.MX7D: Code clean up.
Per systemd.offline-update(7) documentation, the symlink `/system-update'
must exist prior to an offline update and removed early in the
update process to prevent a reboot cycle in the event of a failure.

But if recovery is initiated from U-Boot (e.g., if rootfs is not
bootable), U-Boot is presently unable to create symlinks on ext4 file
systems. So to support this case, the `/system-update' symlink should
never be removed from the recovery file system.